requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
cycler==0.11.0
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.13.1
dslib==0.1.0
fonttools==4.28.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
imutils==0.5.4
incremental==21.3.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0rc1
names==0.3.0
nltk==3.6.5
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/SODIQ/500_project/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
service-identity==21.1.0
setuptools-scm==6.3.2
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
threaded==4.1.0
tomli==1.2.2
tqdm==4.62.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
zope.interface==5.4.0
